Use Modular Pattern
- Add vendor libraries
  - clipboard.min.js
  - jquery.min.js
  - log4javascript.min.js
  - require.min.js
- Use requireJS
- Add app module
- Add random utility
- Add UI module
