+++ index.html
@@ -5,14 +5,18 @@
   <meta charset="UTF-8">
   <title>NPM Module List</title>
   <link rel="stylesheet" href="css/style.css">
-  <script src="//cdnjs.cloudflare.com/ajax/libs/jquery/1.11.3/jquery.min.js"></script>
-  <script src="//cdnjs.cloudflare.com/ajax/libs/log4javascript/1.4.9/log4javascript.min.js"></script>
-  <script src="js/script.js"></script>
 </head>
 
 <body>
 
-  <script src="//cdn.rawgit.com/zenorocha/clipboard.js/master/dist/clipboard.min.js"></script>
+  <!-- <script data-main="js/main" src="//cdnjs.cloudflare.com/ajax/libs/require.js/2.1.20/require.min.js"></script> -->
+  <script>
+    if (typeof requirejs === 'undefined') {
+      document.write('<script data-main="js/main" src="js/vendor/require.min.js"><\/script>');
+    }
+
+  </script>
+
 </body>
 
 </html>
